Update README.md
I aim to create an interactive generative art project: with each click, it will randomly generate different portraits of transgender individuals. When clicking on each small avatar, a dialogue box will appear, and the voice will narrate the personal experiences of transgender individuals. Through this project, I hope to provide people with an opportunity to see and hear their voices.